Create backend for playing cards
