+++ digital-deck-server/models/session.js
@@ -1,5 +1,4 @@
-var cardDistService = require("./cardDistService.js");
-var cardDrawService = require("./cardDrawService.js");
+var cardDistService = require('../services/cardDistService');
 var Player = require("./player.js");
 
 /**
@@ -26,10 +25,5 @@ module.exports = class Session {
     this.players = assignments.players;
     this.table = assignments.table;
     this.deck = assignments.deck;
-  }
-
-  drawCards(playerId, numOfCards) {
-    let player = this.players.find(p => p.playerId === playerId);
-    return cardDrawService.drawCards(this.deck, numOfCards, player);
   }
 };
